Add an option to never jump at all
This is useful if you want to run tests from the SUT and just want to
see possible error messages.

A nice alternative is available, if you have test and sut buffers next
to each other: Jump to the window with the tests, run MakeGreen(), turn
on cursorline, and return.  Or similar procedures.