Add a flag to always use the two main error bars
The new flag 'm' always displays errors with the main two error bars
even for target errors in different buffers.

In combination with `J`, this allows you to use the different-buffer
bars exclusively as a jump notification.  For example, if you execute
tests from the SUT without jumping you already know that the error will
be in a different buffer.