Trigger QuickFixCmdPost/Pre
It looks like those autocommands are not triggered when 'make' is
executed as part of an 'execute' command.

This will trigger QuickFixCmdPre immediately before executing 'make' and
QuickFixCmdPost immediately after opening the quickfix window on
a qf-tab.  The command name is set to 'MakeGreen'

You can use the QuickFixCmdPost event to rearrange the windows or
install mappings into a quickfix buffer.